etl_service/requirements.txt
elasticsearch==8.12.1
python-dotenv==0.21.0
#psycopg2==2.9.9
psycopg2_binary==2.9.9
dataclasses==0.6
pydantic==2.6.3
pydantic-settings==2.2.1
backoff==2.2.1